Fixing unit test and bingogame typo.
